requirements.txt
fastapi[standard]~=0.115.12
uvicorn~=0.34.0
sqlalchemy~=2.0.39
alembic~=1.15.1
python-dotenv~=1.0.1
passlib[bcrypt]~=1.7.4
fastapi-jwt-auth~=0.5.0
python-multipart~=0.0.20